feature: Detect ecosystem mismatches (#45)
Ecosystem language and version is added to both lock file manifests and
to environment state files. This allows detecting when configuration is
changed and there are mismatches between either state and manifest, or
between config and state.

When a mismatch is discovered, the existing environment is deleted and
rebuilt. When a mismatch between manifest and config is discovered, this
requires running `goose upgrade`.

This breaks existing configurations, and requires rebuilding
configuration as well as manually deleting existing environments.