feature: Introduce augmented working directory
This fixes an issue where the containerized invocation resulted in
calculating the same environment path wherever it was invoked from. The
environment var `GOOSE_AUGMENTED_CWD` is introduced and allows passing
in the working directory in the host system to the containerized app so
that it can calculate the same unique per-project environment path as
the non-containerized app.